faidlog-complete-refactor
