+++ demos/adult-subgroupdiscovery/README.md
@@ -1,0 +1,5 @@
+# Fair Subgroup Discovery
+
+These notebooks are obtained from https://github.com/MaurizioPulizzi/fairsd/tree/main/notebooks.
+
+We demonstrated applying this method to a tabular dataset (Adult) and logging the experiment metadata using FAID.
