+++ faid/data/entity-lookup.csv
@@ -1,0 +1,9 @@
+is_array, model, model_p, data, data_p, risk, risk_p, usecase, usecase_p, transparency, transparency_p
+1, data, model_parameters,,,,,,,,,
+0, description, data, summary,,,,,,,,
+0, link, data, dataset_link,,,,,,,,
+0, sensitive, data, sensitivity_types,,,,,,,,
+0, sensitive, data, sensitivity_types,,,,,,,,
+1, risks, considerations, risks,, risks,,,,,,
+0, name, risks, name,risks, description,risks,,,,,
+0, mitigation_strategy, risks, mitigation_strategy, risks, mitigation,risks,,,,,
