+++ faid/logging/conversion.py
@@ -1,0 +1,47 @@
+def sync_risk_to_model():
+    # Implement the logic to sync risk to model
+    pass
+
+def sync_data_to_model():
+    # Implement the logic to sync data to model
+    pass
+
+def sync_usecase_to_model():
+    # Implement the logic to sync usecase to model
+    pass
+
+def sync_model_to_risk():
+    # Implement the logic to sync model to risk
+    pass
+
+def sync_model_to_data():
+    # Implement the logic to sync model to data
+    pass
+
+def sync_model_to_usecase():
+    # Implement the logic to sync model to usecase
+    pass
+
+def sync_risk_to_data():
+    # Implement the logic to sync risk to data
+    pass
+
+def sync_usecase_to_data():
+    # Implement the logic to sync usecase to data
+    pass
+
+def sync_risk_to_transparency():
+    # Implement the logic to sync risk to transparency
+    pass
+
+def sync_data_to_transparency():
+    # Implement the logic to sync data to transparency
+    pass
+
+def sync_model_to_transparency():
+    # Implement the logic to sync model to transparency
+    pass
+
+def sync_usecase_to_transparency():
+    # Implement the logic to sync usecase to transparency
+    pass
