+++ faid/report/__init__.py
@@ -1,0 +1,29 @@
+from faid.report.file_utils import (
+    get_faid_report_folder
+)
+
+from faid.report.report_utils import (
+    generate_data_card_report, 
+    generate_model_card_report, 
+    generate_risk_register_report, 
+    generate_experiment_overview_report, 
+    generate_transparency_report,
+    generate_all_reports
+)
+
+from faid.report.viz_utils import (
+    figure_to_base64str,
+    OntologyChart
+)
+
+__all__ = [
+    'get_faid_report_folder',
+    'generate_data_card_report',
+    'generate_model_card_report',
+    'generate_risk_register_report',
+    'generate_experiment_overview_report',
+    'generate_transparency_report',
+    'generate_all_reports',
+    'figure_to_base64str',
+    'OntologyChart'
+]
