+++ requirements.txt
@@ -14,6 +14,7 @@ amqp==5.2.0
 angle-emb==0.4.6
 aniso8601==9.0.1
 annotated-types==0.7.0
+annoy==1.17.3
 antlr4-python3-runtime==4.9.3
 anyio==4.4.0
 appdirs==1.4.4
@@ -56,6 +57,7 @@ colorlog==6.8.2
 comm==0.2.2
 confection==0.1.5
 configobj==5.0.8
+contextlib2==21.6.0
 contourpy==1.2.1
 cryptography==42.0.8
 cycler==0.12.1
@@ -118,6 +120,7 @@ google-cloud-bigquery==3.25.0
 google-cloud-core==2.4.1
 google-cloud-resource-manager==1.12.5
 google-cloud-storage==2.18.2
+google-cloud-translate==3.16.0
 google-crc32c==1.6.0
 google-pasta==0.2.0
 google-resumable-media==2.7.2
@@ -144,9 +147,11 @@ hyperparameter-tuning==0.3.1
 idna==3.7
 ijson==3.3.0
 ImageHash==4.3.1
+imageio==2.35.1
 immutabledict==4.2.0
 importlib_metadata==7.1.0
 importlib_resources==6.4.0
+iniconfig==2.0.0
 inquirer==3.3.0
 inspect_ai==0.3.38
 ipykernel==6.29.4
@@ -180,10 +185,14 @@ langdetect==1.0.9
 langid==1.1.6
 langsmith==0.0.87
 language_data==1.2.0
+lazy_loader==0.4
+Levenshtein==0.26.0
 libclang==18.1.1
 lightning-utilities==0.11.6
+lit-nlp==1.2
 llm-comparator==0.1
 llvmlite==0.43.0
+lxml==5.3.0
 makefun==1.15.4
 Mako==1.3.5
 marisa-trie==1.2.0
@@ -197,6 +206,7 @@ mdurl==0.1.2
 millify==0.1.1
 mistune==3.0.2
 mixpanel==4.10.1
+ml-collections==0.1.1
 ml-dtypes==0.4.0
 mlcroissant==1.0.7
 mlflow==2.13.2
@@ -217,7 +227,7 @@ networkx==3.3
 nltk==3.8.1
 num2words==0.5.13
 numba==0.60.0
-numpy==1.23.5
+numpy==1.26.4
 omegaconf==2.3.0
 openpyxl==3.1.5
 opentelemetry-api==1.25.0
@@ -239,7 +249,10 @@ phik==0.12.4
 pillow==10.3.0
 platformdirs==3.11.0
 plotly==5.23.0
+pluggy==1.5.0
 ply==3.11
+portalocker==2.10.1
+portpicker==1.6.0
 posthog==2.5.0
 preshed==3.0.9
 prettytable==3.10.2
@@ -261,7 +274,9 @@ pydot==2.0.0
 pygit2==1.15.0
 Pygments==2.18.0
 pygtrie==2.5.0
+pynndescent==0.5.13
 pyparsing==3.1.2
+pytest==8.3.3
 python-dateutil==2.9.0.post0
 python-decouple==3.8
 python-dotenv==1.0.1
@@ -279,6 +294,7 @@ regex==2024.5.15
 requests==2.32.2
 requests-toolbelt==1.0.0
 rich==13.7.1
+rouge-score==0.1.2
 rpds-py==0.20.0
 rsa==4.9
 ruamel.yaml==0.18.6
@@ -286,9 +302,12 @@ ruamel.yaml.clib==0.2.8
 runs==1.2.2
 s3fs==2024.3.0
 s3transfer==0.10.1
+sacrebleu==2.4.3
 sacremoses==0.1.1
 safetensors==0.4.3
+saliency==0.2.1
 schema==0.7.7
+scikit-image==0.24.0
 scikit-learn==1.5.0
 scipy==1.11.4
 scmrepo==3.3.5
@@ -298,12 +317,14 @@ semver==3.0.2
 sentencepiece==0.2.0
 sentry-sdk==2.5.1
 setproctitle==1.3.3
+shap==0.45.1
 shapely==2.0.6
 shellingham==1.5.4
 shortuuid==1.0.13
 shtab==1.7.1
 simple_parsing==0.1.5
 six==1.16.0
+slicer==0.0.8
 smart-open==7.0.4
 smmap==5.0.1
 sniffio==1.3.1
@@ -335,9 +356,11 @@ tf_keras==2.17.0
 thefuzz==0.22.1
 thinc==8.2.5
 threadpoolctl==3.5.0
+tifffile==2024.9.20
 tiktoken==0.7.0
 tokenizers==0.19.1
 toml==0.10.2
+tomli==2.0.2
 tomlkit==0.12.5
 toolz==0.12.1
 torch==2.3.1
@@ -360,6 +383,7 @@ tyro==0.8.4
 tzdata==2024.1
 tzlocal==5.2
 ucimlrepo==0.0.7
+umap-learn==0.5.6
 urllib3==2.2.1
 validators==0.33.0
 vertexai==1.66.0
@@ -376,6 +400,7 @@ Werkzeug==3.0.3
 widgetsnbextension==4.0.11
 wordcloud==1.9.3
 wrapt==1.16.0
+xlrd==2.0.1
 xmod==1.8.1
 xxhash==3.4.1
 yarl==1.9.4
